+++ source/extensions/omni.flux.asset_importer.widget/omni/flux/asset_importer/widget/tests/__init__.py
@@ -15,8 +15,10 @@
 * limitations under the License.
 """
 
+from .e2e.common.test_ingestion_checker import TestIngestionCheckerE2E
 from .e2e.file_import_list.test_file_import_list_widget import TestFileImportListWidget
 from .e2e.texture_import_list.test_texture_import_list_widget import TestTextureImportListWidget
+from .unit.common.test_ingestion_checker import TestIngestionCheckerUnit
 from .unit.file_import_list.test_file_import_list_items import TestFileImportListItems
 from .unit.file_import_list.test_file_import_list_model import TestFileImportListModel
 from .unit.texture_import_list.test_texture_import_list_items import TestTextureImportListItems