+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - REMIX-2874: Added a scan folder dialog for importing
 - REMIX-3443: Implemented USD Visibility Plugins
 - REMIX-2518: Added external asset prevention/copying functionality
+- REMIX-2907: Added warning when invalid file types are dropped for ingestion
 
 ### Changed
 - REMIX-3401: Changed all usages of factories to use the Centralized Generic factory