+++ source/extensions/omni.flux.tabbed.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.4]
+### Fixed
+- Fixed initial tab enabled state
+
 ## [1.1.3]
 ### Changed
 - Update deps